# Performance: Only get and release thread safe context within upsertCompaction() if the serie has compaction rules.